Move expval mitigation folder
