+++ qiskit/ignis/mitigation/expval/__init__.py
@@ -1,0 +1,27 @@
+# -*- coding: utf-8 -*-
+
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2019.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+"""
+Measurement error mitigation module
+"""
+
+# Legacy Measurement Error mitigation
+from .legacy import (complete_meas_cal, tensored_meas_cal, MeasurementFilter,
+                     TensoredFilter, CompleteMeasFitter, TensoredMeasFitter)
+
+from .meas_mit_generator import MeasMitigatorGenerator
+from .meas_mit_fitter import MeasMitigatorFitter
+from .complete_method import CompleteMeasMitigator
+from .tensored_method import TensoredMeasMitigator
+from .ctmp_method import CTMPMeasMitigator
+from .meas_mit_utils import expectation_value, counts_probability_vector
