--- qiskit/ignis/mitigation/measurement/legacy/__init__.py
@@ -1,24 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# This code is part of Qiskit.
-#
-# (C) Copyright IBM 2019.
-#
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
-
-
-"""
-Measurement correction module
-
-"""
-
-# Measurement correction functions
-from .circuits import complete_meas_cal, tensored_meas_cal
-from .filters import MeasurementFilter, TensoredFilter
-from .fitters import CompleteMeasFitter, TensoredMeasFitter
